checkpoint for updates to inf-ruby.el
This changeset was begun after the following error, when using
inf-ruby with irb 1.2.6 (Ruby for Windows) in a GNU Emacs w64 build

Exact Emacs version:
GNU Emacs 27.1 (build 1, x86_64-w64-mingw32) of 2020-08-21

~~~~
irb(main):009:0> frob = [1,"two",3.3,-4]
Errno::EBADF (Bad file descriptor)
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/input-method.rb:42:in `winsize'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/input-method.rb:42:in `winsize'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:753:in `output_value'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:545:in `block (2 levels) in eval_input'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:704:in `signal_status'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:538:in `block in eval_input'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/ruby-lex.rb:166:in `block (2 levels) in each_top_level_statement'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/ruby-lex.rb:151:in `loop'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/ruby-lex.rb:151:in `block in each_top_level_statement'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/ruby-lex.rb:150:in `catch'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/ruby-lex.rb:150:in `each_top_level_statement'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:537:in `eval_input'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:472:in `block in run'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:471:in `catch'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:471:in `run'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:400:in `start'
Maybe IRB bug!
irb(main):009:0> frob = [1,"two",3.3,-4]
Errno::EBADF (Bad file descriptor)
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/input-method.rb:42:in `winsize'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/input-method.rb:42:in `winsize'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:753:in `output_value'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:545:in `block (2 levels) in eval_input'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:704:in `signal_status'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:538:in `block in eval_input'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/ruby-lex.rb:166:in `block (2 levels) in each_top_level_statement'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/ruby-lex.rb:151:in `loop'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/ruby-lex.rb:151:in `block in each_top_level_statement'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/ruby-lex.rb:150:in `catch'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb/ruby-lex.rb:150:in `each_top_level_statement'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:537:in `eval_input'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:472:in `block in run'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:471:in `catch'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:471:in `run'
	from C:/ruby/Ruby27-x64/lib/ruby/2.7.0/irb.rb:400:in `start'
Maybe IRB bug!
~~~~

The goal was to try running irb via the Ruby cygwin build. This has been
successful, under the following configuration. There are nonetheless
some prevailing concerns in the patched code

~~~~
(unless cygwinp
  (setq inf-ruby-subshell "bash"
        inf-ruby-subshell-args '("-c" "'%s'")))
~~~~

The resulting inf-ruby process is produced as completely absent of any
irb prompt string.

It may be recommended to enable such as the following configuration,
when testing this patched code:

~~~~
(unless cygwinp
  (setq explicit-shell-file-name "c:/cygwin64/bin/bash.exe"
        explicit-bash-args '("-l" "-i")))
~~~~

The bad file descriptor error may appear to occur, with or without that
addtional configuration.

Considering the availability of the Emacs explicit-shell-file-name
variable, it may be hypothesized that Emacs will then call irb via the
Cygwin BASH shell. As such, the added inf-ruby-subshell handling may
then be superfluous.

If Emacs is already calling BASH for process exec, it may in effect be
trimming some of the options off if the following shell command is run
via Emacs:
   bash -c 'irb ...'

Moreover, the use of the additional bash subshell may causing some
concerns with regards to pty allocation under Cygwin, such that the
normal irb prompt may not then be available.

In a configuration with this patched code, it has served to prevent the
bad file descriptor error, as seen with the other Ruby build in this
Emacs configuration. However,  the concern about the irb prompt and the
overall design of the patched code may bear some further consideration.

A later update should provide some means to utilize a configuration such
as the following - principally, as to run the ruby interpreter directly
via the subshell in Emacs, ensuring then that the interpreter will load
the appropriate irb script, directly.

~~~~
(unless cygwinp
  (setq ;; inf-ruby-subshell nil ;; n/a after undoing some of this patched code
        inf-ruby-ruby-irb-command "c:/cygwin64/bin/ruby.exe"
        inf-ruby-ruby-irb-args '("/bin/irb" ...)))
~~~~

This forked branch, as such, should be considered to be still in
development.